refactor: move types and constants out of TablePagination component gf-23
