+++ apps/frontend/src/libs/components/table-pagination/libs/constants/constants.ts
@@ -1,0 +1,3 @@
+export { FIRST_PAGE } from "./first-page.constant.js";
+export { PAGE_INCREMENT } from "./page-increment.constant.js";
+export { options as PAGE_SIZE_OPTIONS } from "./page-size-options.constant.js";
