+++ apps/frontend/src/libs/components/table-pagination/libs/constants/first-page.constant.ts
@@ -1,0 +1,3 @@
+const FIRST_PAGE = 1;
+
+export { FIRST_PAGE };
