+++ apps/frontend/src/libs/components/table-pagination/libs/constants/page-increment.constant.ts
@@ -1,0 +1,3 @@
+const PAGE_INCREMENT = 1;
+
+export { PAGE_INCREMENT };
