+++ apps/frontend/src/libs/components/table-pagination/libs/constants/page-size-options.constant.ts
@@ -1,0 +1,9 @@
+import { type PageSizeOption } from "../types/types.js";
+
+const options: PageSizeOption[] = [
+	{ label: "10", value: 10 },
+	{ label: "25", value: 25 },
+	{ label: "50", value: 50 },
+];
+
+export { options };
