+++ apps/frontend/src/libs/components/table-pagination/libs/types/form-data.type.ts
@@ -1,0 +1,7 @@
+import { type PageSizeOption } from "./page-size-option.type.js";
+
+type FormData = {
+	pageSize: PageSizeOption;
+};
+
+export { type FormData };
