+++ apps/frontend/src/libs/components/table-pagination/libs/types/page-size-option.type.ts
@@ -1,0 +1,5 @@
+import { type SelectOption } from "~/libs/types/types.js";
+
+type PageSizeOption = SelectOption<number>;
+
+export { type PageSizeOption };
