+++ apps/frontend/src/libs/components/table-pagination/libs/types/types.ts
@@ -1,0 +1,2 @@
+export { type FormData } from "./form-data.type.js";
+export { type PageSizeOption } from "./page-size-option.type.js";
